// src/app/interfaces/iguard-json.ts
export interface IGuardJson{
    guardId: number;// "101";
    manager: string ;// "100";
    name:string;//"שי מוקד אמון";
    address?: string;// אלכסנדר ינאי 1,
    phone?: string;
    descr?:  string ;
    

    background:string;//"white";
    textColor: string;//"black";
};